:recycle: refactor: divide main logic into each file
